enemy health, damage per minuut 10% increase

+++ shooter/Assets/Scripts/Player/Weapons/WeaponReference.cs
@@ -49,7 +49,7 @@ public class WeaponReference : MonoBehaviour
         float doubleChance = Random.Range(1, 5);
         if(doubleChance == list.itemQuantity[4])
         {
-            Fire2(dir + (0.1f * dir));
+            Fire2((int)dir + (0.1f * dir));
         }
     }
